引入RocketMQ削峰 避免请求直接打到db 完成下单接口

+++ src/main/java/com/miaosha/ordercenter/model/PromoModel.java
@@ -30,7 +30,7 @@ public class PromoModel {
     private Integer itemId;
 
     //商品的活动价
-    private BigDecimal promoItemPrice;
+    private Double promoItemPrice;
 
     //秒杀活动状态，1为未开始，2为进行中，3为已结束
     private Integer promoStatus;
